+++ Stripe/StripeiOSTests/STPImageLibraryTests.swift
@@ -11,7 +11,7 @@ import XCTest
 @testable@_spi(STP) import Stripe
 @testable@_spi(STP) import StripeCore
 @testable@_spi(STP) import StripePayments
-@testable@_spi(STP) @_spi(AppearanceAPIAdditionsPreview) import StripePaymentSheet
+@testable@_spi(ReactNativeSDK) @_spi(AppearanceAPIAdditionsPreview) import StripePaymentSheet
 @testable@_spi(STP) import StripePaymentsUI
 @testable@_spi(STP) import StripeUICore