+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/API Bindings/Link/STPAPIClient+Link.swift
@@ -539,30 +539,23 @@ extension STPAPIClient {
             "request_surface": requestSurface.rawValue,
         ]
 
-        if let details = updateParams.details, case .card(let expiryDate, let billingDetails, let preferredNetwork) = details {
-            if let expiryDate {
-                parameters["exp_month"] = expiryDate.month
-                parameters["exp_year"] = expiryDate.year
-            }
-
-            if let billingDetails = billingDetails {
-                parameters["billing_address"] = billingDetails.consumersAPIParams
-            }
-
-            if let billingEmailAddress = billingDetails?.email, !billingEmailAddress.isEmpty {
-                // This email address needs to be lowercase or the API will reject it
-                parameters["billing_email_address"] = billingEmailAddress.lowercased()
-            }
-
-            if let preferredNetwork {
-                parameters["preferred_network"] = preferredNetwork
+        if let metadata = updateParams.metadata {
+            switch metadata {
+            case .card(let expiryDate, let preferredNetwork):
+                if let expiryDate {
+                    parameters["exp_month"] = expiryDate.month
+                    parameters["exp_year"] = expiryDate.year
+                }
+                if let preferredNetwork {
+                    parameters["preferred_network"] = preferredNetwork
+                }
             }
         }
 
-        if let details = updateParams.details, case .bankAccount(let billingDetails) = details {
+        if let billingDetails = updateParams.billingDetails {
             parameters["billing_address"] = billingDetails.consumersAPIParams
 
-            if let billingEmailAddress = billingDetails.email {
+            if let billingEmailAddress = billingDetails.email, !billingEmailAddress.isEmpty {
                 // This email address needs to be lowercase or the API will reject it
                 parameters["billing_email_address"] = billingEmailAddress.lowercased()
             }